requirements.txt
boto3==1.35.63
fastapi==0.115.8
pydantic==1.10.2
python-dotenv==1.0.1
requests==2.32.3
